+++ src/widgets/index.ts
@@ -5,8 +5,10 @@ import { Expanded } from "./expanded.js";
 import { Widget } from "./framework.js";
 import { IconThemeData } from "./icon_theme_data.js";
 import { Padding } from "./padding.js";
+import { Positioned } from "./positioned.js";
 import { Row } from "./row.js";
 import { Spacer } from "./spacer.js";
+import { Stack } from "./stack.js";
 import { Text } from "./Text.js";
 
 export{
@@ -19,5 +21,7 @@ export{
     Padding,
     Widget,
     IconThemeData,
-    Spacer
+    Spacer,
+    Stack,
+    Positioned
 }