+++ src/windiapp/scaffold.ts
@@ -35,12 +35,13 @@ function Scaffold(args: {
       backgroundColor: Colors.Gray[100].value,
       width: "100%",
       height: "100%",
+      overflow:"hidden"
     }),
   });
 
   if (args.body && args.appBar) {
     widget.tag!.style.paddingTop =
-      args.appBar.tag!.offsetHeight + 20 + "px" ?? "100px";
+      args.appBar.tag!.offsetHeight +  "px" ?? "100px";
     widget.tag!.setAttribute("component", "body");
   }